+++ Source/SwiftLintFramework/File+SwiftLint.swift
@@ -16,26 +16,23 @@ extension File {
         withSyntaxKinds syntaxKinds: [SyntaxKind]) -> [NSRange] {
         return matchPattern(pattern).filter { _, kindsInRange in
             return kindsInRange.count == syntaxKinds.count &&
-                filter(zip(kindsInRange, syntaxKinds), { $0.0 != $0.1 }).count == 0
+                zip(kindsInRange, syntaxKinds).filter({ $0.0 != $0.1 }).count == 0
         }.map { $0.0 }
     }
 
     public func matchPattern(pattern: String) -> [(NSRange, [SyntaxKind])] {
-        return flatMap(NSRegularExpression(pattern: pattern, options: nil, error: nil)) { regex in
-            let range = NSRange(location: 0, length: count(contents.utf16))
-            let syntax = syntaxMap
-            let matches = regex.matchesInString(contents, options: nil, range: range)
-            return map(matches as? [NSTextCheckingResult]) { matches in
-                return matches.map { match in
-                    let tokensInRange = syntax.tokens.filter {
-                        NSLocationInRange($0.offset, match.range)
-                    }
-                    let kindsInRange = compact(map(tokensInRange) {
-                        SyntaxKind(rawValue: $0.type)
-                    })
-                    return (match.range, kindsInRange)
-                }
+        let regex = try! NSRegularExpression(pattern: pattern, options: [])
+        let range = NSRange(location: 0, length: contents.utf16.count)
+        let syntax = syntaxMap
+        let matches = regex.matchesInString(contents, options: [], range: range)
+        return matches.map { match in
+            let tokensInRange = syntax.tokens.filter {
+                NSLocationInRange($0.offset, match.range)
             }
-        } ?? []
+            let kindsInRange = tokensInRange.flatMap {
+                SyntaxKind(rawValue: $0.type)
+            }
+            return (match.range, kindsInRange)
+        }
     }
 }