+++ Source/SwiftLintFramework/Rule.swift
@@ -18,3 +18,22 @@ public protocol ParameterizedRule: Rule {
     typealias ParameterType
     var parameters: [RuleParameter<ParameterType>] { get }
 }
+
+public let allRules: [Rule] = [
+    LineLengthRule(),
+    LeadingWhitespaceRule(),
+    TrailingWhitespaceRule(),
+    ReturnArrowWhitespaceRule(),
+    TrailingNewlineRule(),
+    OperatorFunctionWhitespaceRule(),
+    ForceCastRule(),
+    FileLengthRule(),
+    TodoRule(),
+    ColonRule(),
+    TypeNameRule(),
+    VariableNameRule(),
+    TypeBodyLengthRule(),
+    FunctionBodyLengthRule(),
+    NestingRule(),
+    ControlStatementRule()
+]