+++ Source/SwiftLintFramework/Rules/ReturnArrowWhitespaceRule.swift
@@ -6,6 +6,7 @@
 //  Copyright (c) 2015 Realm. All rights reserved.
 //
 
+import Foundation
 import SourceKittenFramework
 
 public struct ReturnArrowWhitespaceRule: Rule {