+++ Source/SwiftLintFramework/Rules/TrailingNewlineRule.swift
@@ -15,13 +15,13 @@ public struct TrailingNewlineRule: Rule {
 
     public func validateFile(file: File) -> [StyleViolation] {
         let string = file.contents
-        let start = advance(string.endIndex, -2, string.startIndex)
+        let start = string.endIndex.advancedBy(-2, limit: string.startIndex)
         let range = Range(start: start, end: string.endIndex)
         let substring = string[range].utf16
         let newLineSet = NSCharacterSet.newlineCharacterSet()
-        let slices = split(substring, allowEmptySlices: true) { !newLineSet.characterIsMember($0) }
+        let slices = substring.split(allowEmptySlices: true) { !newLineSet.characterIsMember($0) }
 
-        if let slice = slices.last where count(slice) != 1 {
+        if let slice = slices.last where slice.count != 1 {
             return [StyleViolation(type: .TrailingNewline,
                 location: Location(file: file.path, line: file.contents.lines().count + 1),
                 severity: .Medium,