+++ Source/swiftlint/LintCommand.swift
@@ -8,7 +8,7 @@
 
 import Commandant
 import Foundation
-import LlamaKit
+import Result
 import SourceKittenFramework
 import SwiftLintFramework
 
@@ -21,38 +21,44 @@ struct LintCommand: CommandType {
 
     func run(mode: CommandMode) -> Result<(), CommandantError<()>> {
         return LintOptions.evaluate(mode).flatMap { options in
+            let configuration = Configuration(path: options.configurationFile,
+                optional: !Process.arguments.contains("--config"))
             if options.useSTDIN {
                 let standardInput = NSFileHandle.fileHandleWithStandardInput()
                 let stdinData = standardInput.readDataToEndOfFile()
                 let stdinNSString = NSString(data: stdinData, encoding: NSUTF8StringEncoding)
                 if let stdinString = stdinNSString as? String {
-                    let violations = Linter(file: File(contents: stdinString)).styleViolations
-                    println(join("\n", violations.map { $0.description }))
-                    return success()
+                    let file = File(contents: stdinString)
+                    let linter = Linter(file: file, configuration: configuration)
+                    print(linter.styleViolations.map({ $0.description }).joinWithSeparator("\n"))
+                    return .Success()
                 }
-                return failure(CommandantError<()>.CommandError(Box()))
+                return .Failure(CommandantError<()>.CommandError())
             }
 
             // Otherwise parse path.
-            return self.lint(options.path)
+            return lint(options.path, configuration: configuration)
         }
     }
 
-    private func lint(path: String) -> Result<(), CommandantError<()>> {
-        let filesToLint = filesToLintAtPath(path)
+    private func lint(path: String,
+        configuration: Configuration) -> Result<(), CommandantError<()>> {
+        let filesToLint = (configuration.included.count == 0 ? filesToLintAtPath(path) : [])
+            .filter({ !configuration.excluded.flatMap(filesToLintAtPath).contains($0) }) +
+            configuration.included.flatMap(filesToLintAtPath)
         if filesToLint.count > 0 {
-
-            if path == "" {
-                println("Linting Swift files in current working directory")
+            if path.isEmpty {
+                print("Linting Swift files in current working directory")
             } else {
-                println("Linting Swift files at path \(path)")
+                print("Linting Swift files at path \(path)")
             }
-
             var numberOfViolations = 0, numberOfSeriousViolations = 0
-            for (index, file) in enumerate(filesToLint) {
-                println("Linting '\(file.lastPathComponent)' (\(index + 1)/\(filesToLint.count))")
-                for violation in Linter(file: File(path: file)!).styleViolations {
-                    println(violation)
+            for (index, path) in filesToLint.enumerate() {
+                let filename = (path as NSString).lastPathComponent
+                print("Linting '\(filename)' (\(index + 1)/\(filesToLint.count))")
+                let file = File(path: path)!
+                for violation in Linter(file: file, configuration: configuration).styleViolations {
+                    print(violation)
                     numberOfViolations++
                     if violation.severity.isError {
                         numberOfSeriousViolations++
@@ -61,27 +67,23 @@ struct LintCommand: CommandType {
             }
             let violationSuffix = (numberOfViolations != 1 ? "s" : "")
             let filesSuffix = (filesToLint.count != 1 ? "s." : ".")
-            println(
+            print(
                 "Done linting!" +
                 " Found \(numberOfViolations) violation\(violationSuffix)," +
                 " \(numberOfSeriousViolations) serious" +
                 " in \(filesToLint.count) file\(filesSuffix)"
             )
             if numberOfSeriousViolations <= 0 {
-                return success()
-            } else {
-                // This represents failure of the content (i.e. violations in the files linted)
-                // and not failure of the scanning process itself. The current command architecture
-                // doesn't discriminate between these types.
-                return failure(CommandantError<()>.CommandError(Box()))
+                return .Success()
             }
+            return .Failure(CommandantError<()>.CommandError())
         }
-        return failure(CommandantError<()>.UsageError(description: "No lintable files found at" +
+        return .Failure(CommandantError<()>.UsageError(description: "No lintable files found at" +
             " path \(path)"))
     }
 
     private func filesToLintAtPath(path: String) -> [String] {
-        let absolutePath = path.absolutePathRepresentation()
+        let absolutePath = (path.absolutePathRepresentation() as NSString).stringByStandardizingPath
         var isDirectory: ObjCBool = false
         if fileManager.fileExistsAtPath(absolutePath, isDirectory: &isDirectory) {
             if isDirectory {
@@ -99,15 +101,22 @@ struct LintCommand: CommandType {
 struct LintOptions: OptionsType {
     let path: String
     let useSTDIN: Bool
+    let configurationFile: String
 
-    static func create(path: String)(useSTDIN: Bool) -> LintOptions {
-        return LintOptions(path: path, useSTDIN: useSTDIN)
+    static func create(path: String)(useSTDIN: Bool)(configurationFile: String) -> LintOptions {
+        return LintOptions(path: path, useSTDIN: useSTDIN, configurationFile: configurationFile)
     }
 
     static func evaluate(m: CommandMode) -> Result<LintOptions, CommandantError<()>> {
         return create
-            <*> m <| Option(key: "path", defaultValue: "", usage: "the path to the file or" +
-                        " directory to lint")
-            <*> m <| Option(key: "use-stdin", defaultValue: false, usage: "lint standard input")
+            <*> m <| Option(key: "path",
+                defaultValue: "",
+                usage: "the path to the file or directory to lint")
+            <*> m <| Option(key: "use-stdin",
+                defaultValue: false,
+                usage: "lint standard input")
+            <*> m <| Option(key: "config",
+                defaultValue: ".swiftlint.yml",
+                usage: "the path to SwiftLint's configuration file")
     }
 }