+++ Source/swiftlint/StructuredText.swift
@@ -19,23 +19,23 @@ enum StructuredText {
     var markdown: String {
         switch self {
         case let .Header(level, t):
-            return join("", Array(count: level, repeatedValue: "#")) + " \(t)"
+            return String(Repeat(count: level, repeatedValue: "#")) + " \(t)"
         case .Paragraph(let t):
             return t
         case .List(let items):
-            return "\n".join(items.map { "* " + $0.markdown })
+            return items.map({ "* " + $0.markdown }).joinWithSeparator("\n")
         case .Joined(let items):
-            return "\n\n".join(items.map { $0.markdown } )
+            return items.map({ $0.markdown }).joinWithSeparator("\n\n")
         }
     }
 
     var ansi: String {
         switch self {
         case .Header(1, let t): return t.uppercaseString
-        case let .Header(other, t): return t
+        case .Header(_, let t): return t
         case .Paragraph(let t): return t
-        case .List(let items): return "\n".join(items.map { "* " + $0.ansi })
-        case .Joined(let items): return "\n\n".join(items.map { $0.ansi } )
+        case .List(let items): return items.map({ "* " + $0.ansi }).joinWithSeparator("\n")
+        case .Joined(let items): return items.map({ $0.ansi }).joinWithSeparator("\n\n")
         }
     }
 }